feat(reimbursements): AR models take the PORO names; POROs move under Airtable
The boundary POROs (Person, Budget, BudgetForecast, Expense, Batch,
EusaActual) move to Reimbursements::Airtable::* — only the Mapper and tests
construct them, and their unit tests move with them. ActiveRecord models now
occupy the old names with the same public interface: record_id = id.to_s,
string batch_id/budget_id overrides, receipts wrapped into Attachment POROs
over ActiveStorage, Person delegating bank fields to PaymentDetails, and
Budget computing the old Airtable rollups (committed/paid/forecast/remaining/
variance) from expenses + forecasts. New models: PaymentDetails, BudgetOwner,
FinancialYear. The effective-payee money path is a shared EffectivePayee
mixin so both backends compute BACS rows identically during the cutover
window. All new models are excluded from the permission grid (portal-managed).
