+++ app/controllers/admin/permissions_controller.rb
@@ -82,7 +82,10 @@ class Admin::PermissionsController < AdminController
       "Event" => { "add_non_members" => "Add non-members to events, mainly for archiving purposes" }
     }
 
-    @models = (ApplicationRecord.descendants + [ Admin::Debt, Season, Doorkeeper::Application ] - [ MarketingCreatives::CategoryInfo, Admin::Proposals::Proposal, OpportunityRole, Reimbursements::BatchAttempt, Reimbursements::OwnerEndorsement ]).uniq
+    @models = (ApplicationRecord.descendants + [ Admin::Debt, Season, Doorkeeper::Application ] - [ MarketingCreatives::CategoryInfo, Admin::Proposals::Proposal, OpportunityRole, Reimbursements::BatchAttempt, Reimbursements::OwnerEndorsement,
+                  Reimbursements::Person, Reimbursements::PaymentDetails, Reimbursements::Budget,
+                  Reimbursements::BudgetOwner, Reimbursements::BudgetForecast, Reimbursements::Expense,
+                  Reimbursements::Batch, Reimbursements::EusaActual, Reimbursements::FinancialYear ]).uniq
 
     role_exclude = Admin::Permission::EXCLUDED_ROLES
     @roles = Role.includes(:permissions).where.not(name: role_exclude).all.left_joins(:permissions).group(:id).order("COUNT(admin_permissions.id) DESC")
