+++ app/models/reimbursements/budget_forecast.rb
@@ -1,22 +1,46 @@
+# == Schema Information
+#
+# Table name: reimbursements_budget_forecasts
+# Database name: primary
+#
+#  id                 :bigint           not null, primary key
+#  amount             :decimal(12, 2)
+#  date               :date
+#  reason             :text(65535)
+#  created_at         :datetime         not null
+#  updated_at         :datetime         not null
+#  airtable_record_id :string(255)
+#  budget_id          :bigint           not null
+#
+# Indexes
+#
+#  index_reimbursements_budget_forecasts_on_airtable_record_id  (airtable_record_id) UNIQUE
+#  index_reimbursements_budget_forecasts_on_budget_id           (budget_id)
+#
+# Foreign Keys
+#
+#  fk_rails_...  (budget_id => reimbursements_budgets.id)
+#
 module Reimbursements
   ##
-  # A versioned projected-expenditure update for a Budget, stored in the
-  # Airtable Budget Forecasts table. Each row is one dated projection with a
-  # reason; the Budgets table rolls the latest up into +current_forecast+. The
-  # finance Budgets screen shows a budget's forecast history and lets the
-  # operator append a new one.
-  class BudgetForecast
-    attr_reader :record_id, :budget_id, :amount, :date, :reason, :name
+  # A versioned projected-expenditure update for a budget. The latest row
+  # (date desc) is the budget's current_forecast. ActiveRecord replacement for
+  # the Airtable-era PORO (now Reimbursements::Airtable::BudgetForecast).
+  class BudgetForecast < ApplicationRecord
+    belongs_to :budget, class_name: "Reimbursements::Budget", inverse_of: :forecasts
 
-    # +amount+ is a BigDecimal (or nil); +date+ a Date (or nil). +name+ is an
-    # Airtable formula label, read-only.
-    def initialize(record_id:, budget_id: nil, amount: nil, date: nil, reason: "", name: "")
-      @record_id = record_id
-      @budget_id = budget_id
-      @amount = amount
-      @date = date
-      @reason = reason
-      @name = name
+    validates :amount, presence: true
+
+    def record_id = id&.to_s
+
+    # The PORO exposed the linked budget's record id string (compared against
+    # budget.record_id in the Store and views); AR's own reader would return
+    # the integer FK.
+    def budget_id = self[:budget_id]&.to_s
+
+    # The Airtable "Name" formula label: "<budget> - YYYY-MM-DD".
+    def name
+      [ budget&.name, date&.strftime("%Y-%m-%d") ].compact.join(" - ")
     end
   end
 end
